+++ pkg/builder/registration_builder.go
@@ -1,0 +1,72 @@
+package builder
+
+import (
+	"github.com/turbonomic/turbo-go-sdk/pkg/proto"
+)
+
+
+
+//// Helper methods to create AccountDefinition map for sub classes of the probe
+// An AccountDefEntryBuilder builds an AccountDefEntry instance.
+type AccountDefEntryBuilder struct {
+	accountDefEntry *proto.AccountDefEntry
+}
+
+func NewAccountDefEntryBuilder(name, displayName, description, verificationRegex string,
+				mandatory  bool,	//proto.AccountDefEntry_AccountDefEntryType,
+				isSecret bool) *AccountDefEntryBuilder {
+	fieldType := &proto.CustomAccountDefEntry_PrimitiveValue_{
+		PrimitiveValue: proto.CustomAccountDefEntry_STRING,
+	}
+	entry := &proto.CustomAccountDefEntry {
+		Name: &name,
+		DisplayName: &displayName,
+		Description: &description,
+		VerificationRegex: &verificationRegex,
+		IsSecret: &isSecret,
+		FieldType: fieldType,
+	}
+
+	customDef := &proto.AccountDefEntry_CustomDefinition {
+		CustomDefinition: entry,
+	}
+
+	accountDefEntry := &proto.AccountDefEntry{
+		Mandatory: &mandatory,
+		Definition: customDef,
+	}
+
+	return &AccountDefEntryBuilder{
+		accountDefEntry: accountDefEntry,
+	}
+}
+
+func (builder *AccountDefEntryBuilder) Create() *proto.AccountDefEntry {
+	return builder.accountDefEntry
+}
+
+// A ProbeInfoBuilder builds a ProbeInfo instance.
+type ProbeInfoBuilder struct {
+	probeInfo *proto.ProbeInfo
+}
+
+// NewProbeInfoBuilder builds the ProbeInfo DTO for the given probe
+func NewProbeInfoBuilder(probeType, probeCat string,
+				supplyChainSet []*proto.TemplateDTO,
+				acctDef []*proto.AccountDefEntry) *ProbeInfoBuilder {
+	// New ProbeInfo protobuf with this input
+	probeInfo := &proto.ProbeInfo{
+		ProbeType:                &probeType,
+		ProbeCategory:            &probeCat,
+		SupplyChainDefinitionSet: supplyChainSet,
+		AccountDefinition:        acctDef,
+	}
+	return &ProbeInfoBuilder{
+		probeInfo: probeInfo,
+	}
+}
+
+func (builder *ProbeInfoBuilder) Create() *proto.ProbeInfo {
+	return builder.probeInfo
+}
+
