--- pkg/communication/MessageOperation.go
@@ -1,1 +1,0 @@
-package communication
