+++ pkg/communication/client_protobuf_endpoint.go
