+++ pkg/communication/client_websocket_transport.go
