+++ pkg/communication/protobuf_endpoint.go
