--- pkg/probe/IDispatch.go
@@ -1,6 +1,0 @@
-package probe
-
-type IDispatch interface {
-	deployEntity()
-	addTarget()
-}
