+++ pkg/service/tap_service.go
@@ -7,6 +7,9 @@ import (
 	"os"
 	"encoding/json"
 
+	restclient "github.com/turbonomic/turbo-api/pkg/client"
+	vmtapi "github.com/turbonomic/turbo-go-sdk/pkg/vmtapi"
+
 	"github.com/turbonomic/turbo-go-sdk/pkg/communication"
 	"github.com/turbonomic/turbo-go-sdk/pkg/probe"
 )
@@ -15,17 +18,28 @@ type TAPService struct {
 	// Interface to the Turbo Server
 	*communication.MediationContainer
 	*probe.TurboProbe
-	*TurboAPIHandler
-	IsRegistered		chan bool
+	*vmtapi.TurboAPIHandler	// TODO: use vmtapi.Client
+	*restclient.Client
+}
+
+func (tapService *TAPService) ConnectToTurbo() {
+	IsRegistered := make(chan bool, 1)
+
+	// Connect to the Turbo server
+	go tapService.MediationContainer.Init(IsRegistered)
+
+	// start a separate go routine to listen for probe registration and create targets in turbo server
+	tapService.createTurboTargets(IsRegistered)
 }
 
 // Invokes the Turbo Rest API to create VMTTarget representing the target environment
 // that is being controlled by the TAP service.
 // Targets are created only after the service is notified of successful registration with the server
-func (tapService *TAPService) addTargets()  {
+func (tapService *TAPService) createTurboTargets(IsRegistered chan bool)  {
+	fmt.Printf("[******************* TAPService *****************] Waiting for registration complete .... %s\n", IsRegistered)
 	// Block till a message arrives on the channel
-	isRegistered := <- tapService.IsRegistered
-	if !isRegistered {
+	status := <- IsRegistered
+	if !status {
 		fmt.Println("[TAPService] Probe " + tapService.ProbeCategory + "::" + tapService.ProbeType + " should be registered before adding Targets")
 		return
 	}
@@ -33,28 +47,51 @@ func (tapService *TAPService) addTargets()  {
 	var targets []*probe.TurboTarget
 	targets = tapService.GetProbeTargets()
 	for _, targetInfo := range targets {
-		tapService.AddTarget(targetInfo)
+		fmt.Println("[TAPService] Adding target %s", targetInfo)
+		tapService.AddTurboTarget(targetInfo)
+		// TODO: use this api call
+		//targetData := &vmttype.Target{
+		//	Category: tapService.ProbeCategory,
+		//	Type:     targetInfo.GetTargetType(),
+		//	InputFields: []*vmttype.InputField{
+		//		{
+		//			Value:           &targetInfo.GetNameOrAddress(),
+		//			Name:            "nameOrAddress",
+		//			GroupProperties: []*vmttype.List{},
+		//		},
+		//		{
+		//			Value:           "username",
+		//			Name:            "VC-username",
+		//			GroupProperties: []*vmttype.List{},
+		//		},
+		//		{
+		//			Value:           "password",
+		//			Name:            "VC-password",
+		//			GroupProperties: []*vmttype.List{},
+		//		},
+		//	},
+		//}
+		//tapService.AddTarget(targetData)
 	}
 }
 
-func (tapService *TAPService) ConnectToTurbo() {
-	// Connect to the Turbo server
-	tapService.MediationContainer.Init(tapService.IsRegistered)
-
-	// start a separate go routine to listen for probe registration and create targets in turbo server
-	go tapService.addTargets()
-}
 
 // ==============================================================================
 
+// Configuration parameters for communicating with the Turbo server
 type TurboCommunicationConfig struct  {
-	*TurboAPIConfig
+	// Config for the Rest API client
+	*vmtapi.TurboAPIConfig
+	// Config for RemoteMediation client that communicates using websocket
 	*communication.ContainerConfig
 }
 
 func parseTurboCommunicationConfig (configFile string) *TurboCommunicationConfig {
 	// load the config
 	turboCommConfig := readTurboCommunicationConfig (configFile)
+	if turboCommConfig == nil {
+		os.Exit(1)
+	}
 	fmt.Println("WebscoketContainer Config : ", turboCommConfig.ContainerConfig)
 	fmt.Println("RestAPI Config: ", turboCommConfig.TurboAPIConfig)
 
@@ -62,6 +99,7 @@ func parseTurboCommunicationConfig (configFile string) *TurboCommunicationConfig
 	// TODO: return validation errors
 	turboCommConfig.ValidateContainerConfig()
 	turboCommConfig.ValidateTurboAPIConfig()
+	fmt.Println("---------- Loaded Turbo Communication Config ---------")
 	return turboCommConfig
 }
 
@@ -78,20 +116,23 @@ func readTurboCommunicationConfig (path string) *TurboCommunicationConfig {
 
 	if err != nil {
 		fmt.Printf("[TurboCommunicationConfig] Unmarshall error :%v\n", err)
+		return nil
 	}
 	fmt.Printf("[TurboCommunicationConfig] Results: %+v\n", config)
 	return &config
 }
 
 // ==============================================================================
+// Convenience builder for building a TAPService
 type TAPServiceBuilder struct {
 	tapService *TAPService
 }
-// Get an instance of ClientMessageBuilder
+
+// Get an instance of TAPServiceBuilder
 func NewTAPServiceBuilder () *TAPServiceBuilder {
 	serviceBuilder := &TAPServiceBuilder{}
 	service := &TAPService {
-		IsRegistered: make(chan bool, 1),	// buffered channel so the send does not block
+		//IsRegistered: make(chan bool, 1),	// buffered channel so the send does not block
 	}
 	serviceBuilder.tapService = service
 	return serviceBuilder
@@ -108,18 +149,31 @@ func (pb *TAPServiceBuilder) Create() *TAPService {
 }
 
 // The Communication Layer to communicate with the Turbo server
+// Uses Websocket to listen to server requests for discovery and actions
+// Uses Rest API to send requests to server for deployment
 func (pb *TAPServiceBuilder) WithTurboCommunicator(commConfFile string) *TAPServiceBuilder {
-	//  Load the main container configuration file and validate it
+	//  Load the main communication configuration file and validate it
 	fmt.Println("[TAPServiceBuilder] TurboCommunicator configuration from %s", commConfFile)
 	commConfig := parseTurboCommunicationConfig(commConfFile)
-	fmt.Println("---------- Loaded Turbo Communication Config ---------")
+
 	// The Webscoket Container
 	theContainer := communication.CreateMediationContainer(commConfig.ContainerConfig)
 	pb.tapService.MediationContainer = theContainer
 	// The RestAPI Handler
 	// TODO: if rest api config has validation errors or not specified, do not create the handler
-	turboApiHandler := NewTurboAPIHandler(commConfig.TurboAPIConfig)
+	turboApiHandler := vmtapi.NewTurboAPIHandler(commConfig.TurboAPIConfig)
 	pb.tapService.TurboAPIHandler = turboApiHandler
+
+	//config := restclient.NewConfigBuilder(commConfig.TurboAPIConfig.VmtRestServerAddress).
+	//	APIPath("/vmturbo/rest").
+	//	BasicAuthentication(commConfig.TurboAPIConfig.VmtRestUser, commConfig.TurboAPIConfig.VmtRestPassword).//"<UI-username>", "UI-password").
+	//	Create()
+	//client, err := restclient.NewAPIClientWithBA(config)
+	//if err != nil {
+	//	fmt.Errorf("[TAPServiceBuilder] Error creating  Turbo Rest API client: %s", err)
+	//}
+	//pb.tapService.Client = client
+
 	return pb
 }
